feat(app.js): add MongoDB connection and error handling
The MongoDB connection has been added to the app.js file. The connection string is set to "mongodb://127.0.0.1:27017/express-mongoose-recipes-dev". The connection is established using the mongoose library. If the connection is successful, a success message is logged to the console, including the name of the connected database. If there is an error connecting to MongoDB, an error message is logged to the console.